+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Version 1.2
 
+## Version 1.2.2
+  - Add `pdiff` function with signature `pdiff(f, x, y,..., 1)` for the partial derivative of `f(x, y,...)` with respect to first argument, namely `x`.
+  - Add `curl` function that accepts 3-dimensional vector fields `F(x, y, z)` with signature `curl(F, x, y, z)`
+  - Add `div` function that accepts vector fields `F(x, y, ...)` with signature `div(F, x, y, ...)`
+
 ## Version 1.2.1
   - fixed a bug that caused some successfully loaded graphs to crash when new object added (#200)
   - added inverse trig functions `arctan`, `arcsin`, `arccos`. The `arctan` function can be called with 1 or 2 arguments.